requirements.txt
kivy==2.2.1
kivymd==1.1.1
opencv-python==4.8.1.78
pyzbar==0.1.9
Pillow==10.0.1
requests==2.31.0
numpy==1.24.3
openai-whisper==20231117

